useract.h, c/c.ast, c/c_type.{cc,h}: Replace 'long' with '[u]intptr_t'.
In those places where I was using an integer to hold a pointer, use
the standard type.  This is important on Windows since it is IL32P64.

+++ c/c_type.cc
@@ -226,7 +226,7 @@ NamedAtomicType::~NamedAtomicType()
 string NamedAtomicType::uniqueName() const
 {
   // 'a' for atomic
-  return stringc << "a" << (long)this /*id*/ << "_" << name;
+  return stringc << "a" << (intptr_t)this /*id*/ << "_" << name;
 }
